refactor(pr03): G3b PhysicalSystemData.mts -> PhysicalItemSystemData.mts
- G3b.1 git mv PhysicalSystemData.mts -> PhysicalItemSystemData.mts (Windows two-step)

- G3b.2 update sibling index.mts and PhysicalItemSystemModel.mts import paths

- G3b.3 npm run build clean

Pure file rename: the file holds 'PhysicalItemSystemData' and 'PhysicalItemSystemSource' interfaces, so file name should match per R.2.5.
