+++ src/entities/items/components/Physical/data/PhysicalItemSystemData.mts
